requirements.txt
aiohttp==3.6.2
aioredis==1.3.1
asgiref==2.3.2
async-timeout==3.0.0
attrs==20.1.0
autobahn==20.7.1
Automat==20.2.0
beautifulsoup4==4.9.2
certifi==2020.6.20
cffi==1.14.2
channels==2.1.2
channels-redis==2.3.2
chardet==3.0.4
constantly==15.1.0
cryptography==3.0
daphne==2.2.0
defusedxml==0.6.0
dj-stripe==2.3.0
Django==2.2
django-allauth==0.32.0
django-bootstrap-modal-forms==2.0.0
django-countries==6.1.3
django-crispy-forms==1.2.4
django-paypal==1.0.0
django-widget-tweaks==1.4.8
ffprobe==0.5
geoip2==4.1.0
google==3.0.0
gunicorn==20.0.4
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
idna-ssl==1.1.0
incremental==17.5.0
jsonfield==3.1.0
lxml==4.5.2
maxminddb==2.0.2
msgpack==0.5.2
msgpack-python==0.5.6
multidict==4.7.6
oauthlib==3.1.0
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.0
paypalrestsdk==1.13.1
Pillow==7.2.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydub==0.24.1
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python3-openid==3.2.0
pytz==2020.1
redis==2.10.6
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
stripe==2.54.0
Twisted==20.3.0
txaio==20.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
websocket-client==0.57.0
yarl==1.5.1
zope.interface==5.1.0
